timely-backend/requirements.txt
channels==4.3.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
stripe==12.5.1
